Fixed new fields mapping into entities

# Conflicts:
#	mysql/db/from_step_05.sql
#	pom.xml
#	rsc/database.properties
#	src/org/isf/admission/service/AdmissionIoOperations.java
#	src/org/isf/menu/service/MenuIoOperations.java

+++ src/org/isf/lab/model/LaboratoryRow.java
@@ -101,7 +101,7 @@ public class LaboratoryRow extends Auditable<String>
 	        final int m = 23;
 	        int c = 133;
 	        
-	        c = m * c + (code == null ? 0 : code.intValue());
+	        c = m * c + (code == null ? 0 : code);
 	        
 	        this.hashCode = c;
 	    }
